Separating queries and presentational components
