+++ src/types/BlogPost.ts
@@ -1,0 +1,48 @@
+export interface INode {
+  parent: {
+    changeTime: Date
+    relativeDirectory: string
+  }
+  body: string
+  id: string
+  frontmatter: {
+    title: string
+    date: Date
+    update: Date
+    tags: string[]
+    spoiler: string
+    link?: string
+    score?: number
+    image: {
+      childImageSharp: {
+        fluid: any
+      }
+    }
+  }
+}
+
+export interface IPageContext {
+  prev: {
+    parent: {
+      relativeDirectory: string
+    }
+    frontmatter: {
+      title: string
+      tags: string[]
+    }
+  }
+  next: {
+    parent: {
+      relativeDirectory: string
+    }
+    frontmatter: {
+      title: string
+      tags: string[]
+    }
+  }
+}
+
+export interface BlogPostProps {
+  pageData: INode
+  pageContext: IPageContext
+}
